Bugfix to use numpad as well as number row
pressing a number on the number row and pressing a command like D to draw, you would draw that number of cards.

This did not work with the numpad.

With this change it works with the numpad. it works by instead of looking at the ID of the key pressed (old: number 0 maps to 48 in number row only), it reads the actual key being pressed (new: number 0 maps to 0 anywhere).

had to do a for loop since setting a range with >= and <= was allowing letters to be passed through which would crash with int64()